requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
executing==2.2.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langchainhub==0.1.21
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.4
langsmith==0.4.21
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.102.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
tavily-python==0.7.11
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
